chore: move utils & constants
