+++ src/infrastructure/bg_tasks/preprocess_dataset/dependencies.py
@@ -2,5 +2,5 @@ from src.infrastructure.storage.client import S3Storage
 from src.infrastructure.storage.client import get_storage as create_storage
 
 
-async def get_storage() -> S3Storage:
+def get_storage() -> S3Storage:
     return create_storage()
